+++ src/platforms/Ayrshare/AsTikTok.ts
@@ -1,7 +1,7 @@
-import Ayrshare from "./Ayrshare";
-import { PlatformId } from ".";
-import Folder from "../core/Folder";
-import Post from "../core/Post";
+import Ayrshare from "../Ayrshare";
+import { PlatformId } from "..";
+import Folder from "../../core/Folder";
+import Post from "../../core/Post";
 
 /**
  * AsTikTok: support for tiktok platform through Ayrshare
